rename server to server-graphql
