+++ packages/e2e/__tests__/password.ts
@@ -1,4 +1,4 @@
-import { ServerTest } from './servers/server';
+import { ServerTest } from './servers/server-graphql';
 
 const server = new ServerTest();
